chore: refactor types model
