+++ services/cover/src/service/build_config.rs
@@ -1,7 +1,8 @@
 use crate::common::types::CallerId;
 use crate::service::model::build_config::BuildConfig;
+use crate::service::model::error::Error;
 use crate::service::{build_config, build_config_mut};
-use crate::{CanisterId, Error};
+use crate::CanisterId;
 
 pub fn get_all_build_configs(caller_id: &CallerId) -> Vec<&'static BuildConfig> {
     build_config().get_all_build_configs(caller_id)
@@ -15,6 +16,7 @@ pub fn get_build_config_by_id(
         .get_build_config_by_id(caller_id, canister_id)
         .map_err(|e| e.into())
 }
+
 pub fn update_build_config(
     caller_id: &CallerId,
     canister_id: &CanisterId,
