+++ services/cover/src/service/model/error.rs
@@ -1,0 +1,9 @@
+use ic_kit::candid::CandidType;
+use serde::Deserialize;
+
+#[derive(CandidType, Deserialize)]
+pub struct Error {
+    pub code: &'static str,
+    pub message: &'static str,
+    pub debug_log: Option<String>,
+}
