+++ services/cover/src/service/model/mod.rs
@@ -1,1 +1,6 @@
 pub mod build_config;
+pub mod error;
+pub mod progress;
+pub mod provider;
+pub mod request;
+pub mod verification;
