+++ services/cover/src/service/store/test_data/provider_info.rs
@@ -1,4 +1,4 @@
-use crate::service::types::ProviderInfo;
+use crate::service::model::provider::ProviderInfo;
 
 pub fn fake_provider_info1() -> ProviderInfo {
     ProviderInfo {}
