+++ services/cover/src/service/store/verification.rs
@@ -5,9 +5,9 @@ use ic_kit::candid::CandidType;
 use serde::Deserialize;
 
 use crate::common::types::{CallerId, CanisterId};
+use crate::service::model::verification::{AddVerification, UpdateVerification, Verification};
 use crate::service::store::error::ErrorKindStore;
 use crate::service::time_utils;
-use crate::service::types::{AddVerification, UpdateVerification, Verification};
 
 #[derive(CandidType, Deserialize, Default)]
 pub struct VerificationStore {
